// src/shared/types/action-type.ts
export enum ActionType {
  TRANSACTION_APPROVE = "TRANSACTION_APPROVE",
  TRANSACTION_REJECT = "TRANSACTION_REJECT",
  DEPOSIT_CONFIRM = "DEPOSIT_CONFIRM",
  DOCUMENT_APPROVE = "DOCUMENT_APPROVE",
  DOCUMENT_REJECT = "DOCUMENT_REJECT",
  COMPLIANCE_REVIEW = "COMPLIANCE_REVIEW",
  USER_SUSPEND = "USER_SUSPEND",
  USER_ACTIVATE = "USER_ACTIVATE",
  LIMITS_OVERRIDE = "LIMITS_OVERRIDE",
  CASH_PICKUP_ISSUE = "CASH_PICKUP_ISSUE",
  PREPAID_CARD_ISSUE = "PREPAID_CARD_ISSUE",
  AGENT_CREATE = "AGENT_CREATE",
  AGENT_UPDATE_STATUS = "AGENT_UPDATE_STATUS",
  AGENT_APPROVE = "AGENT_APPROVE",
  AGENT_DEACTIVATE = "AGENT_DEACTIVATE",
  RATE_CREATE = "RATE_CREATE",
  RATE_UPDATE = "RATE_UPDATE",
  RATE_DEACTIVATE = "RATE_DEACTIVATE",
  REPORT_GENERATE = "REPORT_GENERATE",
  ADMIN_USER_CREATE = "ADMIN_USER_CREATE",
  ADMIN_USER_UPDATE = "ADMIN_USER_UPDATE",
  ADMIN_USER_ACTIVATE = "ADMIN_USER_ACTIVATE",
  ADMIN_USER_SUSPEND = "ADMIN_USER_SUSPEND",
  ROLE_CREATE = "ROLE_CREATE",
  ROLE_UPDATE = "ROLE_UPDATE",
  ROLE_DELETE = "ROLE_DELETE",
  DEPARTMENT_CREATE = "DEPARTMENT_CREATE",
  DEPARTMENT_UPDATE = "DEPARTMENT_UPDATE",
  DEPARTMENT_DELETE = "DEPARTMENT_DELETE",
  TICKET_CREATE = "TICKET_CREATE",
  TICKET_UPDATE = "TICKET_UPDATE",
  TICKET_ASSIGN = "TICKET_ASSIGN",
  TICKET_STATUS_UPDATE = "TICKET_STATUS_UPDATE",
  TICKET_COMMENT = "TICKET_COMMENT",
  WALLET_ADD_NOTE = "WALLET_ADD_NOTE",
  WALLET_LINK_TRANSACTION = "WALLET_LINK_TRANSACTION",
  WALLET_UNLINK_TRANSACTION = "WALLET_UNLINK_TRANSACTION",
  WALLET_FLAG_ENTRY = "WALLET_FLAG_ENTRY",
  WALLET_REFUND = "WALLET_REFUND",
  WALLET_DISBURSE = "WALLET_DISBURSE",
  TRANSACTION_REFUND = "TRANSACTION_REFUND",
  DISBURSEMENT_INITIATE = "DISBURSEMENT_INITIATE",
  DISBURSEMENT_APPROVE = "DISBURSEMENT_APPROVE",
  DISBURSEMENT_REJECT = "DISBURSEMENT_REJECT",
}